@@ -1,18 +1,46 @@
 "use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
+    Object.defineProperty(o, "default", { enumerable: true, value: v });
+}) : function(o, v) {
+    o["default"] = v;
+});
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 Object.defineProperty(exports, "__esModule", { value: true });
-const fs = require("fs");
-const ejs = require("ejs");
-const path = require("path");
+const fs = __importStar(require("fs"));
+const ejs = __importStar(require("ejs"));
+const path = __importStar(require("path"));
 let optionsWhiteList = ['type', 'length', 'primaryKey', 'autoIncrement', 'notNull', 'unique', 'defaultValue', 'foreignKey'];
 class MigrationBuilder {
-    constructor(modelsPrimaryKeysTypes) {
+    constructor(modelsPrimaryKeysTypes, modelsList) {
         this.migrationsBuild = "";
         this.modelsPrimaryKeysTypes = modelsPrimaryKeysTypes;
+        this.modelsList = modelsList;
     }
     createTable(tableName, columnSpec) {
         for (let column in columnSpec) {
             let processedColumnName = this.processColumnName(column, columnSpec[column]);
             columnSpec[column] = this.processColumnSpec(tableName, processedColumnName, columnSpec[column]);
+            // do not create a migration if field is an association
+            if (columnSpec[column] === null) {
+                delete columnSpec[column];
+            }
         }
         this.migrationsBuild = this.migrationsBuild.concat(`db.createTable('${tableName}', {\n` +
             `    columns: ${JSON.stringify(columnSpec)},\n` +
@@ -22,11 +50,19 @@ class MigrationBuilder {
     addColumn(tableName, columnName, columnSpec) {
         let processedColumnName = this.processColumnName(columnName, columnSpec);
         columnSpec = this.processColumnSpec(tableName, columnName, columnSpec);
+        // do not create a migration if field is an association
+        if (columnSpec === null) {
+            return;
+        }
         this.migrationsBuild = this.migrationsBuild.concat(`db.addColumn('${tableName}', '${processedColumnName}', ${JSON.stringify(columnSpec)});\n`);
     }
     changeColumn(tableName, columnName, columnSpec) {
         let processedColumnName = this.processColumnName(columnName, columnSpec);
         columnSpec = this.processColumnSpec(tableName, columnName, columnSpec);
+        // do not create a migration if field is an association
+        if (columnSpec === null) {
+            return;
+        }
         this.migrationsBuild = this.migrationsBuild.concat(`db.changeColumn('${tableName}', '${processedColumnName}', ${JSON.stringify(columnSpec)});\n`);
     }
     dropTable(tableName) {
@@ -46,18 +82,31 @@ class MigrationBuilder {
     processColumnSpec(tableName, columnName, columnSpec) {
         // process collections
         if (columnSpec.collection) {
+            // if (!this.modelsList.includes(columnSpec.collection)) {
+            //   throw `Model ${tableName} has an association to model ${columnSpec.collection}, but model ${columnSpec.collection} is not presented in models tree`;
+            // }
+            // !TODO если нету columnSpec.via, то columnSpec.via = this.modelsPrimaryKeysTypes[columnSpec.collection];
             let tableFieldsType = 'string';
             if (this.modelsPrimaryKeysTypes[columnSpec.collection]) { // fields' types will be like primaryKey in related model
                 tableFieldsType = this.modelsPrimaryKeysTypes[columnSpec.collection];
             }
+            // db-migrate should check if intermediate table exists, then skip creating the table (ejs)
+            // this case is only for outside model collection
             if (!this.migrationsBuild.includes(`${columnSpec.collection}_${columnSpec.via}__${tableName}_${columnName}`)) {
                 this.createTable(`${tableName}_${columnName}__${columnSpec.collection}_${columnSpec.via}`, {
-                    id: { type: 'int', notNull: true },
+                    id: { type: 'int', notNull: true, autoIncrement: true },
                     [`${tableName}_${columnName}`]: { type: tableFieldsType },
                     [`${columnSpec.collection}_${columnSpec.via}`]: { type: tableFieldsType }
                 });
             }
+            return null; // do not create a migration to this field
         }
+        if (columnSpec.model) {
+            return null; // do not create a migration to this field
+        }
+        // if (columnSpec.model && !this.modelsList.includes(columnSpec.model)) {
+        //   throw `Model ${tableName} has an association to model ${columnSpec.model}, but model ${columnSpec.model} is not presented in models tree`;
+        // }
         // process columnSpec options
         for (let key in columnSpec) {
             if (key === 'defaultsTo') {
