@@ -1,21 +1,29 @@
 import * as fs from "fs";
 import * as ejs from "ejs";
 import * as path from "path";
+import Base from "db-migrate-base";
+import {AttributeSpec, ModelSpec} from "../interfaces/types";
 let optionsWhiteList = ['type', 'length', 'primaryKey', 'autoIncrement', 'notNull', 'unique', 'defaultValue', 'foreignKey'];
 
 export default class MigrationBuilder {
-  private migrationsBuild;
-  private modelsPrimaryKeysTypes;
+  private migrationsBuild: string;
+  private readonly modelsPrimaryKeysTypes: {[key:string]: string};
+  private modelsList: string[];
 
-  constructor(modelsPrimaryKeysTypes) {
+  constructor(modelsPrimaryKeysTypes, modelsList) {
     this.migrationsBuild = "";
     this.modelsPrimaryKeysTypes = modelsPrimaryKeysTypes;
+    this.modelsList = modelsList;
   }
 
-  public createTable(tableName, columnSpec) {
+  public createTable(tableName: string, columnSpec: ModelSpec): void {
     for (let column in columnSpec) {
       let processedColumnName = this.processColumnName(column, columnSpec[column]);
       columnSpec[column] = this.processColumnSpec(tableName, processedColumnName, columnSpec[column]);
+      // do not create a migration if field is an association
+      if (columnSpec[column] === null) {
+        delete columnSpec[column]
+      }
     }
     this.migrationsBuild = this.migrationsBuild.concat(`db.createTable('${tableName}', {\n` +
       `    columns: ${JSON.stringify(columnSpec)},\n` +
@@ -23,27 +31,35 @@ export default class MigrationBuilder {
       `  });\n`);
   }
 
-  public addColumn(tableName, columnName, columnSpec) {
+  public addColumn(tableName: string, columnName: string, columnSpec: AttributeSpec): void {
     let processedColumnName = this.processColumnName(columnName, columnSpec);
     columnSpec = this.processColumnSpec(tableName, columnName, columnSpec);
+    // do not create a migration if field is an association
+    if (columnSpec === null) {
+      return
+    }
     this.migrationsBuild = this.migrationsBuild.concat(`db.addColumn('${tableName}', '${processedColumnName}', ${JSON.stringify(columnSpec)});\n`);
   }
 
-  public changeColumn(tableName, columnName, columnSpec) {
+  public changeColumn(tableName: string, columnName: string, columnSpec: AttributeSpec): void {
     let processedColumnName = this.processColumnName(columnName, columnSpec);
     columnSpec = this.processColumnSpec(tableName, columnName, columnSpec);
+    // do not create a migration if field is an association
+    if (columnSpec === null) {
+      return
+    }
     this.migrationsBuild = this.migrationsBuild.concat(`db.changeColumn('${tableName}', '${processedColumnName}', ${JSON.stringify(columnSpec)});\n`);
   }
 
-  public dropTable(tableName) {
+  public dropTable(tableName: string): void {
     this.migrationsBuild = this.migrationsBuild.concat(`db.dropTable('${tableName}');\n`);
   }
 
-  public removeColumn(tableName, columnName) {
+  public removeColumn(tableName: string, columnName: string): void {
     this.migrationsBuild = this.migrationsBuild.concat(`db.removeColumn('${tableName}', '${columnName}');\n`);
   }
 
-  public processColumnName(columnName, columnSpec) {
+  public processColumnName(columnName: string, columnSpec: AttributeSpec): string {
     for (let key in columnSpec) {
       if (key === 'columnName') {
         columnName = columnSpec[key];
@@ -52,21 +68,39 @@ export default class MigrationBuilder {
     return columnName;
   }
 
-  public processColumnSpec(tableName, columnName, columnSpec) {
+  public processColumnSpec(tableName: string, columnName: string, columnSpec: AttributeSpec): Base.ColumnSpec {
     // process collections
     if (columnSpec.collection) {
+      // if (!this.modelsList.includes(columnSpec.collection)) {
+      //   throw `Model ${tableName} has an association to model ${columnSpec.collection}, but model ${columnSpec.collection} is not presented in models tree`;
+      // }
+
+      // !TODO если нету columnSpec.via, то columnSpec.via = this.modelsPrimaryKeysTypes[columnSpec.collection];
+
       let tableFieldsType = 'string';
       if (this.modelsPrimaryKeysTypes[columnSpec.collection]) { // fields' types will be like primaryKey in related model
         tableFieldsType = this.modelsPrimaryKeysTypes[columnSpec.collection];
       }
+      // db-migrate should check if intermediate table exists, then skip creating the table (ejs)
+      // this case is only for outside model collection
       if (!this.migrationsBuild.includes(`${columnSpec.collection}_${columnSpec.via}__${tableName}_${columnName}`)) {
         this.createTable(`${tableName}_${columnName}__${columnSpec.collection}_${columnSpec.via}`, {
-          id: {type: 'int', notNull: true},
+          id: {type: 'int', notNull: true, autoIncrement: true},
           [`${tableName}_${columnName}`]: {type: tableFieldsType},
           [`${columnSpec.collection}_${columnSpec.via}`]: {type: tableFieldsType}
         })
       }
+
+      return null; // do not create a migration to this field
     }
+
+    if (columnSpec.model) {
+      return null; // do not create a migration to this field
+    }
+
+    // if (columnSpec.model && !this.modelsList.includes(columnSpec.model)) {
+    //   throw `Model ${tableName} has an association to model ${columnSpec.model}, but model ${columnSpec.model} is not presented in models tree`;
+    // }
 
     // process columnSpec options
     for (let key in columnSpec) {
@@ -96,10 +130,10 @@ export default class MigrationBuilder {
         columnSpec.type = 'real';
       }
     }
-    return columnSpec
+    return columnSpec as Base.ColumnSpec
   }
 
-  public renderFile() {
+  public renderFile(): void {
     let date = new Date();
     let fileName = `${date.getFullYear()}${pad2(date.getMonth() + 1)}${pad2(date.getDate())}${pad2(date.getHours())}${pad2(date.getMinutes())}${pad2(date.getSeconds())}-${process.env.MIGRATION_NAME}`;
 
@@ -113,11 +147,11 @@ export default class MigrationBuilder {
     })
   }
 
-  public getMigrationsBuild() {
+  public getMigrationsBuild(): string {
     return this.migrationsBuild;
   }
 }
 
-function pad2(n) {
+function pad2(n: number): string {
   return (n < 10 ? '0' : '') + n;
 }
