@@ -2,42 +2,42 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 class DB {
     constructor() {
-        this.waterlineSchema = {};
+        this.intermediateMigrationSchema = {};
     }
-    createTable(tableName, columnSpec, callback) {
+    createTable(tableName, columnSpec) {
         // !TODO process other parameters in columnSpec
-        this.waterlineSchema[tableName] = {};
+        this.intermediateMigrationSchema[tableName] = {};
         let columns = columnSpec.columns ? columnSpec.columns : columnSpec;
         for (let key in columns) {
-            this.waterlineSchema[tableName][key] = this.processColumn(columns[key]);
+            this.intermediateMigrationSchema[tableName][key] = this.processColumn(columns[key]);
         }
     }
-    addColumn(tableName, columnName, columnSpec, callback) {
-        this.waterlineSchema[tableName][columnName] = this.processColumn(columnSpec);
+    addColumn(tableName, columnName, columnSpec) {
+        this.intermediateMigrationSchema[tableName][columnName] = this.processColumn(columnSpec);
     }
-    dropTable(tableName, options, callback) {
+    dropTable(tableName, options) {
         // !TODO how to process this options ?
         // options will be an array, to process this case we need usage example
         // what means [options,] here: dropTable(tableName, [options,] callback) ?
-        delete this.waterlineSchema[tableName];
+        delete this.intermediateMigrationSchema[tableName];
     }
-    renameTable(tableName, newTableName, callback) {
-        this.waterlineSchema[newTableName] = this.waterlineSchema[tableName];
-        delete this.waterlineSchema[tableName];
+    renameTable(tableName, newTableName) {
+        this.intermediateMigrationSchema[newTableName] = this.intermediateMigrationSchema[tableName];
+        delete this.intermediateMigrationSchema[tableName];
     }
-    removeColumn(tableName, columnName, callback) {
-        delete this.waterlineSchema[tableName][columnName];
+    removeColumn(tableName, columnName) {
+        delete this.intermediateMigrationSchema[tableName][columnName];
     }
-    renameColumn(tableName, oldColumnName, newColumnName, callback) {
-        this.waterlineSchema[tableName][newColumnName] = this.waterlineSchema[tableName][oldColumnName];
-        delete this.waterlineSchema[tableName][oldColumnName];
+    renameColumn(tableName, oldColumnName, newColumnName) {
+        this.intermediateMigrationSchema[tableName][newColumnName] = this.intermediateMigrationSchema[tableName][oldColumnName];
+        delete this.intermediateMigrationSchema[tableName][oldColumnName];
     }
-    changeColumn(tableName, columnName, columnSpec, callback) {
-        this.waterlineSchema[tableName][columnName] = this.processColumn(columnSpec);
+    changeColumn(tableName, columnName, columnSpec) {
+        this.intermediateMigrationSchema[tableName][columnName] = this.processColumn(columnSpec);
     }
     // !TODO add all other methods
     getWaterlineSchema() {
-        return this.waterlineSchema;
+        return this.intermediateMigrationSchema;
     }
     processColumn(column) {
         if (column.type === 'char' || column.type === 'text' || column.type === 'date' || column.type === 'datetime' ||
